Destructure state management and data to parent component up state
